% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dc}
\alias{dc}
\title{Interacted data with covariates}
\format{
A data frame with observable variables:
\describe{
\item{group}{Group identifier}
\item{groupW}{Covariate group identifier (e.g., State)}
\item{X}{Treatment variable}
\item{Y}{Outcome variable}
\item{e}{Error term}
\item{MX}{Residuals}
\item{Me}{Residuals}
\item{MY}{Residuals}
}
}
\usage{
dc
}
\description{
A simulated dataset for the QOB example with covariates.
Generated using \code{GenData_cov()}.
}
\keyword{datasets}
