% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{GenData_cov}
\alias{GenData_cov}
\title{Title}
\usage{
GenData_cov(
  S = 5,
  p1 = 7/8,
  Het = 3,
  sigeps = 0.5,
  sigev = 0.1,
  beta = 0,
  beta0 = 0,
  omega = 0.1,
  K = 20,
  c = 5
)
}
\arguments{
\item{c}{}
}
\description{
Title
}
