% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{GetCIcoef_ijloop}
\alias{GetCIcoef_ijloop}
\title{Title}
\usage{
GetCIcoef_ijloop(
  df,
  IdPW,
  IdPQ,
  dPW,
  dPQ,
  X,
  Y,
  MX,
  MY,
  q = qnorm(0.975)^2,
  noisyi = FALSE,
  noisyj = FALSE
)
}
\arguments{
\item{noisyj}{}
}
\description{
Title
}
