% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{GetCIcoef_ijloop_cov}
\alias{GetCIcoef_ijloop_cov}
\title{Title}
\usage{
GetCIcoef_ijloop_cov(
  X,
  Y,
  MX,
  MY,
  W,
  Q,
  WWWinv,
  QQQinv,
  IdPQ,
  IdPW,
  q = qnorm(0.975)^2,
  noisyi = TRUE,
  noisyj = FALSE
)
}
\arguments{
\item{noisyj}{}
}
\description{
Title
}
