% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{GetCIcoef_iloop}
\alias{GetCIcoef_iloop}
\title{Title}
\usage{
GetCIcoef_iloop(
  df,
  P,
  G,
  X,
  Y,
  MX,
  MY,
  Z,
  W,
  q = qnorm(0.975)^2,
  noisy = FALSE
)
}
\arguments{
\item{noisy}{}
}
\description{
Title
}
