% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{GetL3OCIcoef_fast}
\alias{GetL3OCIcoef_fast}
\title{Title}
\usage{
GetL3OCIcoef_fast(
  df,
  groupW,
  groupQ,
  X,
  Y,
  MX,
  MY,
  q = qnorm(0.975)^2,
  noisy = FALSE
)
}
\arguments{
\item{noisy}{}
}
\description{
Title
}
