% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l3o_variance.R
\name{L3Ovar_ijloop_cov}
\alias{L3Ovar_ijloop_cov}
\title{Title}
\usage{
L3Ovar_ijloop_cov(
  X,
  e,
  MX,
  Me,
  W,
  Q,
  WWWinv,
  QQQinv,
  IdPQ,
  IdPW,
  noisyi = TRUE,
  noisyj = FALSE
)
}
\arguments{
\item{noisyj}{}
}
\description{
Title
}
