% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwivhet-package.R
\docType{package}
\name{mwivhet-package}
\alias{mwivhet}
\alias{mwivhet-package}
\title{mwivhet: Inference with Many Weak Instruments and Heterogeneity}
\description{
Calculate test statistics and confidence intervals in a linear instrumental variables model with possibly many weak instruments and heterogeneous treatment effects, using methods developed in Yap (2025) \url{https://arxiv.org/abs/2408.11193}. Implements a novel score statistic and a leave-three-out variance estimator that is asymptotically uniformly most powerful unbiased. Also computes Lagrange Multiplier tests for robust inference in the presence of high-dimensional covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/minhnngo/mwivhet}
  \item Report bugs at \url{https://github.com/minhnngo/mwivhet/issues}
}

}
\author{
\strong{Maintainer}: Ngoc Minh Ngo \email{minh.ngongoc2@gmail.com} (Package implementation and documentation)

Authors:
\itemize{
  \item Ngoc Minh Ngo \email{minh.ngongoc2@gmail.com} (Package implementation and documentation)
  \item Luther Yap \email{lyap@nus.edu.sg} (Original method and code implementation)
}

}
\keyword{internal}
