% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suffolk}
\alias{suffolk}
\title{Nonviolent Misdemeanor Prosecution in Suffolk County}
\format{
A data frame with observable variables
}
\source{
Agan, A., Doleac, J. L., & Harvey, A. (2023). Misdemeanor Prosecution.
The Quarterly Journal of Economics, 138(3), 1453-1505.
}
\usage{
suffolk
}
\description{
A dataset containing administrative records from the Suffolk County District
Attorney's Office (SCDAO) used to estimate the causal effects of
misdemeanor prosecution on recidivism
}
\keyword{datasets}
