% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{GenData_nocov}
\alias{GenData_nocov}
\title{Title}
\usage{
GenData_nocov(
  S = 3,
  Het = 3,
  sigee = 1,
  sigvv = 1,
  sigexi = 0,
  sigev = 0,
  ConTE = FALSE,
  beta = 0,
  beta0 = 0,
  K = 40,
  c = 5
)
}
\arguments{
\item{c}{}
}
\description{
Title
}
